Added camera shake and COD sound effects. Changed some dependencies

+++ Assets/Scripts/Interactions/WallGun.cs
@@ -4,6 +4,6 @@ public class WallGun : MonoBehaviour
 {
     public GunProperties wallGunProperties;
 
-    public float GetGunPrice() { return wallGunProperties.gunPrice; }
+    public int GetGunPrice() { return wallGunProperties.gunPrice; }
     public FPSGeneral.GunType GetGunType() { return wallGunProperties.gunType; }
 }
